+++ packages/jest-worker/src/workers/ChildProcessWorker.ts
@@ -6,6 +6,8 @@
  */
 
 import childProcess, {ChildProcess} from 'child_process';
+import {PassThrough} from 'stream';
+import mergeStream from 'merge-stream';
 import supportsColor from 'supports-color';
 
 import {
@@ -43,12 +45,18 @@ export default class ChildProcessWorker implements WorkerInterface {
   private _child!: ChildProcess;
   private _options: WorkerOptions;
   private _onProcessEnd!: OnEnd;
+  private _fakeStream: PassThrough | null;
   private _request: ChildMessage | null;
   private _retries!: number;
+  private _stderr: ReturnType<typeof mergeStream> | null;
+  private _stdout: ReturnType<typeof mergeStream> | null;
 
   constructor(options: WorkerOptions) {
     this._options = options;
+    this._fakeStream = null;
     this._request = null;
+    this._stderr = null;
+    this._stdout = null;
 
     this.initialize();
   }
@@ -68,6 +76,26 @@ export default class ChildProcessWorker implements WorkerInterface {
       ...this._options.forkOptions,
     });
 
+    if (child.stdout) {
+      if (!this._stdout) {
+        // We need to add a permanent stream to the merged stream to prevent it
+        // from ending when the subprocess stream ends
+        this._stdout = mergeStream(this._getFakeStream());
+      }
+
+      this._stdout.add(child.stdout);
+    }
+
+    if (child.stderr) {
+      if (!this._stderr) {
+        // We need to add a permanent stream to the merged stream to prevent it
+        // from ending when the subprocess stream ends
+        this._stderr = mergeStream(this._getFakeStream());
+      }
+
+      this._stderr.add(child.stderr);
+    }
+
     child.on('message', this.onMessage.bind(this));
     child.on('exit', this.onExit.bind(this));
 
@@ -79,6 +107,7 @@ export default class ChildProcessWorker implements WorkerInterface {
     ]);
 
     this._child = child;
+
     this._retries++;
 
     // If we exceeded the amount of retries, we will emulate an error reply
@@ -94,6 +123,14 @@ export default class ChildProcessWorker implements WorkerInterface {
         error.stack!,
         {type: 'WorkerError'},
       ]);
+    }
+  }
+
+  private _shutdown() {
+    // End the temporary streams so the merged streams end too
+    if (this._fakeStream) {
+      this._fakeStream.end();
+      this._fakeStream = null;
     }
   }
 
@@ -149,6 +186,8 @@ export default class ChildProcessWorker implements WorkerInterface {
       if (this._request) {
         this._child.send(this._request);
       }
+    } else {
+      this._shutdown();
     }
   }
 
@@ -166,25 +205,22 @@ export default class ChildProcessWorker implements WorkerInterface {
     this._child.send(request);
   }
 
-  getWorkerId(): number {
+  getWorkerId() {
     return this._options.workerId;
   }
 
-  getStdout(): NodeJS.ReadableStream {
-    if (this._child.stdout === null) {
-      throw new Error(
-        'stdout is null - this should never happen. Please open up an issue at https://github.com/facebook/jest',
-      );
-    }
-    return this._child.stdout;
+  getStdout(): NodeJS.ReadableStream | null {
+    return this._stdout;
   }
 
-  getStderr(): NodeJS.ReadableStream {
-    if (this._child.stderr === null) {
-      throw new Error(
-        'stderr is null - this should never happen. Please open up an issue at https://github.com/facebook/jest',
-      );
+  getStderr(): NodeJS.ReadableStream | null {
+    return this._stderr;
+  }
+
+  private _getFakeStream() {
+    if (!this._fakeStream) {
+      this._fakeStream = new PassThrough();
     }
-    return this._child.stderr;
+    return this._fakeStream;
   }
 }