+++ packages/jest-worker/src/workers/__tests__/NodeThreadsWorker.test.js
@@ -9,6 +9,8 @@
 
 /* eslint-disable no-new */
 
+import getStream from 'get-stream';
+
 import {
   CHILD_MESSAGE_CALL,
   CHILD_MESSAGE_INITIALIZE,
@@ -24,10 +26,12 @@ beforeEach(() => {
   jest.mock('worker_threads', () => {
     const fakeClass = jest.fn(() => {
       const EventEmitter = require('events');
+      const {PassThrough} = require('stream');
+
       const thread = new EventEmitter();
       thread.postMessage = jest.fn();
-      thread.stdout = 'stdout';
-      thread.stderr = 'stderr';
+      thread.stdout = new PassThrough();
+      thread.stderr = new PassThrough();
       return thread;
     });
 
@@ -134,15 +138,25 @@ it('stops initializing the worker after the amount of retries is exceeded', () =
   expect(onProcessEnd.mock.calls[0][1]).toBe(null);
 });
 
-it('provides stdout and stderr fields from the child process', () => {
+it('provides stdout and stderr from the child processes', async () => {
   const worker = new Worker({
     forkOptions: {},
     maxRetries: 3,
     workerPath: '/tmp/foo',
   });
 
-  expect(worker.getStdout()).toBe('stdout');
-  expect(worker.getStderr()).toBe('stderr');
+  const stdout = worker.getStdout();
+  const stderr = worker.getStderr();
+
+  worker._worker.stdout.end('Hello ', {encoding: 'utf8'});
+  worker._worker.stderr.end('Jest ', {encoding: 'utf8'});
+  worker._worker.emit('exit');
+  worker._worker.stdout.end('World!', {encoding: 'utf8'});
+  worker._worker.stderr.end('Workers!', {encoding: 'utf8'});
+  worker._worker.emit('exit', 0);
+
+  await expect(getStream(stdout)).resolves.toEqual('Hello World!');
+  await expect(getStream(stderr)).resolves.toEqual('Jest Workers!');
 });
 
 it('sends the task to the child process', () => {